requirements.txt
PyQt5>=5.15.9
cryptography>=41.0.0
Pillow>=10.0.0
requests>=2.31.0